+++ dwds/test/frontend_server_evaluate_test.dart
@@ -77,6 +77,9 @@ void main() async {
       Isolate isolate;
       ScriptList scripts;
       ScriptRef mainScript;
+      ScriptRef libraryScript;
+      ScriptRef testLibraryScript;
+      ScriptRef testLibraryPartScript;
       Stream<Event> stream;
 
       setUp(() async {
@@ -89,6 +92,12 @@ void main() async {
 
         mainScript = scripts.scripts
             .firstWhere((each) => each.uri.contains('main.dart'));
+        testLibraryScript = scripts.scripts.firstWhere((each) =>
+            each.uri.contains('package:_testPackage/test_library.dart'));
+        testLibraryPartScript = scripts.scripts.firstWhere((each) =>
+            each.uri.contains('package:_testPackage/src/test_part.dart'));
+        libraryScript = scripts.scripts.firstWhere(
+            (each) => each.uri.contains('package:_test/library.dart'));
       });
 
       tearDown(() async {
@@ -107,6 +116,90 @@ void main() async {
               result,
               const TypeMatcher<InstanceRef>().having(
                   (instance) => instance.valueAsString, 'valueAsString', '42'));
+        });
+      });
+
+      test('local', () async {
+        await onBreakPoint(isolate.id, mainScript, 'printLocal', () async {
+          var event = await stream
+              .firstWhere((event) => event.kind == EventKind.kPauseBreakpoint);
+
+          var result = await setup.service
+              .evaluateInFrame(isolate.id, event.topFrame.index, 'local');
+
+          expect(
+              result,
+              isA<InstanceRef>().having(
+                  (instance) => instance.valueAsString, 'valueAsString', '42'));
+        });
+      });
+
+      test('field', () async {
+        await onBreakPoint(isolate.id, mainScript, 'printFieldFromLibraryClass',
+            () async {
+          var event = await stream
+              .firstWhere((event) => event.kind == EventKind.kPauseBreakpoint);
+
+          var result = await setup.service.evaluateInFrame(
+              isolate.id, event.topFrame.index, 'instance.field');
+
+          expect(
+              result,
+              isA<InstanceRef>().having(
+                  (instance) => instance.valueAsString, 'valueAsString', '1'));
+        });
+      });
+
+      test('private field from another library', () async {
+        await onBreakPoint(isolate.id, mainScript, 'printFieldFromLibraryClass',
+            () async {
+          var event = await stream
+              .firstWhere((event) => event.kind == EventKind.kPauseBreakpoint);
+
+          var result = await setup.service.evaluateInFrame(
+              isolate.id, event.topFrame.index, 'instance._field');
+
+          expect(
+              result,
+              isA<ErrorRef>().having((instance) => instance.message, 'message',
+                  contains("The getter '_field' isn't defined")));
+        });
+      });
+
+      test('private field from current library', () async {
+        await onBreakPoint(isolate.id, mainScript, 'printFieldMain', () async {
+          var event = await stream
+              .firstWhere((event) => event.kind == EventKind.kPauseBreakpoint);
+
+          var result = await setup.service.evaluateInFrame(
+              isolate.id, event.topFrame.index, 'instance._field');
+
+          expect(
+              result,
+              isA<InstanceRef>().having(
+                  (instance) => instance.valueAsString, 'valueAsString', '1'));
+        });
+      });
+
+      test('access instance fields after evaluation', () async {
+        await onBreakPoint(isolate.id, mainScript, 'printFieldFromLibraryClass',
+            () async {
+          var event = await stream
+              .firstWhere((event) => event.kind == EventKind.kPauseBreakpoint);
+
+          var instanceRef = await setup.service.evaluateInFrame(
+              isolate.id, event.topFrame.index, 'instance') as InstanceRef;
+
+          var instance = await setup.service
+              .getObject(isolate.id, instanceRef.id) as Instance;
+
+          var field = instance.fields
+              .firstWhere((BoundField element) => element.decl.name == 'field');
+
+          expect(
+              field.value,
+              isA<InstanceRef>().having(
+                  (instance) => instance.valueAsString, 'valueAsString', '1'));
         });
       });
 
@@ -172,6 +265,36 @@ void main() async {
         });
       });
 
+      test('call library part function with const param', () async {
+        await onBreakPoint(isolate.id, mainScript, 'printLocal', () async {
+          var event = await stream
+              .firstWhere((event) => event.kind == EventKind.kPauseBreakpoint);
+
+          var result = await setup.service.evaluateInFrame(
+              isolate.id, event.topFrame.index, 'testLibraryPartFunction(42)');
+
+          expect(
+              result,
+              isA<InstanceRef>().having(
+                  (instance) => instance.valueAsString, 'valueAsString', '42'));
+        });
+      });
+
+      test('call library part function with local param', () async {
+        await onBreakPoint(isolate.id, mainScript, 'printLocal', () async {
+          var event = await stream
+              .firstWhere((event) => event.kind == EventKind.kPauseBreakpoint);
+
+          var result = await setup.service.evaluateInFrame(isolate.id,
+              event.topFrame.index, 'testLibraryPartFunction(local)');
+
+          expect(
+              result,
+              isA<InstanceRef>().having(
+                  (instance) => instance.valueAsString, 'valueAsString', '42'));
+        });
+      });
+
       test('loop variable', () async {
         await onBreakPoint(isolate.id, mainScript, 'printLoopVariable',
             () async {
@@ -185,6 +308,87 @@ void main() async {
               result,
               const TypeMatcher<InstanceRef>().having(
                   (instance) => instance.valueAsString, 'valueAsString', '1'));
+        });
+      });
+
+      test('evaluate expression in _testPackage/test_library', () async {
+        await onBreakPoint(isolate.id, testLibraryScript, 'testLibraryFunction',
+            () async {
+          var event = await stream
+              .firstWhere((event) => event.kind == EventKind.kPauseBreakpoint);
+
+          var result = await setup.service
+              .evaluateInFrame(isolate.id, event.topFrame.index, 'formal');
+
+          expect(
+              result,
+              isA<InstanceRef>().having(
+                  (instance) => instance.valueAsString, 'valueAsString', '23'));
+        });
+      });
+
+      test('evaluate expression in a class constructor in a library', () async {
+        await onBreakPoint(
+            isolate.id, testLibraryScript, 'testLibraryClassConstructor',
+            () async {
+          var event = await stream
+              .firstWhere((event) => event.kind == EventKind.kPauseBreakpoint);
+
+          var result = await setup.service
+              .evaluateInFrame(isolate.id, event.topFrame.index, 'this.field');
+
+          expect(
+              result,
+              isA<InstanceRef>().having(
+                  (instance) => instance.valueAsString, 'valueAsString', '1'));
+        });
+      });
+
+      test('evaluate expression in a class constructor in a library part',
+          () async {
+        await onBreakPoint(isolate.id, testLibraryPartScript,
+            'testLibraryPartClassConstructor', () async {
+          var event = await stream
+              .firstWhere((event) => event.kind == EventKind.kPauseBreakpoint);
+
+          var result = await setup.service
+              .evaluateInFrame(isolate.id, event.topFrame.index, 'this.field');
+
+          expect(
+              result,
+              isA<InstanceRef>().having(
+                  (instance) => instance.valueAsString, 'valueAsString', '1'));
+        });
+      });
+
+      test('evaluate expression in caller frame', () async {
+        await onBreakPoint(isolate.id, testLibraryScript, 'testLibraryFunction',
+            () async {
+          var event = await stream
+              .firstWhere((event) => event.kind == EventKind.kPauseBreakpoint);
+
+          var result = await setup.service
+              .evaluateInFrame(isolate.id, event.topFrame.index + 1, 'local');
+
+          expect(
+              result,
+              isA<InstanceRef>().having(
+                  (instance) => instance.valueAsString, 'valueAsString', '23'));
+        });
+      });
+
+      test('evaluate expression in a library', () async {
+        await onBreakPoint(isolate.id, libraryScript, 'Concatenate', () async {
+          var event = await stream
+              .firstWhere((event) => event.kind == EventKind.kPauseBreakpoint);
+
+          var result = await setup.service
+              .evaluateInFrame(isolate.id, event.topFrame.index, 'a');
+
+          expect(
+              result,
+              isA<InstanceRef>().having((instance) => instance.valueAsString,
+                  'valueAsString', 'Hello'));
         });
       });