+++ fixtures/_testPackage/web/main.dart
@@ -24,15 +24,18 @@ void main() {
   Timer.periodic(const Duration(seconds: 1), (_) {
     print('Count is: ${++count}');
     print(testLibraryValue);
+    print(testPartLibraryValue);
   });
 
   // for evaluation
   Timer.periodic(const Duration(seconds: 1), (_) {
     printLocal();
-    printField();
+    printFieldFromLibraryClass();
+    printFieldFromLibraryPartClass();
     printFieldMain();
     printGlobal();
     printFromTestLibrary();
+    printFromTestLibraryPart();
     printFromTestPackage();
     printCallExtension();
     printLoopVariable();
@@ -46,9 +49,14 @@ void printLocal() {
   print('Local is: $local'); // Breakpoint: printLocal
 }
 
-void printField() {
+void printFieldFromLibraryClass() {
   var instance = TestLibraryClass(1, 2);
-  print('$instance'); // Breakpoint: printField
+  print('$instance'); // Breakpoint: printFieldFromLibraryClass
+}
+
+void printFieldFromLibraryPartClass() {
+  var instance = TestLibraryPartClass(1, 2);
+  print('$instance'); // Breakpoint: printFieldFromLibraryPartClass
 }
 
 void printFieldMain() {
@@ -67,6 +75,11 @@ void printFromTestPackage() {
 void printFromTestLibrary() {
   var local = 23;
   print(testLibraryFunction(local));
+}
+
+void printFromTestLibraryPart() {
+  var local = 23;
+  print(testLibraryPartFunction(local));
 }
 
 void printCallExtension() {